% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{aoristos}
\alias{aoristos}
\title{Produce a time series that counts up or otherwise summarises artefact data per year}
\usage{
aoristos(
  startdate,
  enddate,
  weight = NA,
  timeRange = NA,
  calendar = "BCAD",
  verbose = TRUE
)
}
\arguments{
\item{startdate}{A vector of numeric values specifying the approximate start dates for the production of each artefact.}

\item{enddate}{A vector of numeric values specifying the approximate end dates for the production of each artefact.}

\item{weight}{A single numeric vector >0 to be used in weighting the counts or other estimates per year (e.g. object size or weight)..}

\item{timeRange}{Numeric vector of length 2, specifying the earliest and latest date to include in the results (in the year units of the calendar argument). The default option will calculate this automatically from the input data.}

\item{calendar}{Specifies whether the input calendar is in years BC (-ve)/AD (+ve) or BP (default is "BCAD"). Note there should be no year 0 in "BCAD".}

\item{verbose}{A logical variable indicating whether extra information on progress should be reported. Default is TRUE.}
}
\value{
A data.frame with the number artefacts whose production range overlaps with that year, an aoristic sum and the same adjusted for weight.
}
\description{
Given a set of archaeological artefacts with either estimated absolute start and end dates (e.g. based on their style) or a probability model for their likely production intensity through time), count up how many artefacts are definitely or possibly present for each year in a specified time range.
}
\references{
#' Ratcliffe, J.H. 2000. Aoristic analysis: the spatial interpretation of unspecifed temporal events, International Journal of Geographical Information Science 14: 669–679.x#'
Johnson, I. 2004. Aoristic analysis: seeds of a new approach to mapping archaeological distributions through time, In K.F. Ausserer, W. Börner, M. Goriany and L. Karlhuber-Vöckl (eds.), [Enter the past]. The E-way into the Four Dimensions of Cultural Heritage (CAA 2003): 448–452. Oxford: Archaeopress.x#'
}
