% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacities.R
\name{findInterior}
\alias{findInterior}
\title{Find the interior profile of a vessel}
\usage{
findInterior(x, rimoffset = 0, left = TRUE)
}
\arguments{
\item{x}{A vector of numeric values specifying the approximate start dates for the production of each artefact.}

\item{rimoffset}{A single numeric value specifying how much (in the input units) to go down from the tow of the vessel rim before creating the internal shape. Default is 0.}

\item{left}{Logical whether the input and output are the left-hand or right-hand profiles.}
}
\value{
An object of class SpatialPolygons showing the interior left or right hand cavity of a vessel.

#' @examples
data(dolium)
realheight <- 0.72
plot(dolium, col="grey", border=NA, axes=TRUE)
doliumint1 <- findInterior(dolium, rimoffset=realheight/10)
plot(doliumint1, border="red", add=TRUE)
}
\description{
Given a drawn outline of a vessel, find the interior profil only and create a new polygon around the left or right half of the interior cavity.
}
