% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot3dSR}
\alias{plot3dSR}
\title{Find the interior profile of a vessel}
\usage{
plot3dSR(
  x,
  rimoffset = 0,
  left = TRUE,
  interior = TRUE,
  col.interior = "red",
  col.exterior = "darkgoldenrod3",
  windowRect = c(20, 30, 800, 800),
  userMatrix = rotationMatrix(180/pi * 50, 1, 1, 0),
  ...
)
}
\arguments{
\item{x}{An object of class SpatialPolygons* delineating the profile of a vessel.}

\item{rimoffset}{A single numeric value specifying how much (in the input units) to go down from the top of the vessel rim before creating the internal shape. Default is 0.}

\item{left}{Is it the left or the right-hand side profile.}

\item{interior}{Logical for whether to plot the interior cavity as well.}

\item{col.interior}{Colour of the interior}

\item{col.exterior}{Colour of the exterior}

\item{...}{passed to par3d().}
}
\description{
Given a drawn outline of a vessel, find the interior profil only and create a new polygon around the left or right half of the interior cavity.
}
\examples{
\dontrun{
data(dolium)
plot3dSR(dolium)
}

}
