% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacities.R
\name{volsr}
\alias{volsr}
\title{Calculate the volume of a simple solid of revolution}
\usage{
volsr(intprofile, left = TRUE)
}
\arguments{
\item{intprofile}{An object of class SpatialPolygons* delineating the outline of the left- or right-hand side of the interior of an object.}

\item{left}{Is it the left or the right-hand side.}
}
\value{
A single numeric value for the capacity in the cubed units of the input.
}
\description{
Given a left or right-hand half of the interior of a simple (no reversions) vessel shape, calculate the likely capacity if it were swept 360 degres about its middle axis, via the piecewise cylinders methods
}
\examples{
data(dolium)
plot(dolium, col="grey", border=NA, axes=TRUE)
plot(doliumint, border="red", add=TRUE)
volsr(doliumint) ##Input units are metres, capacity is 12.8L
}
