% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capacities.R
\name{volSR}
\alias{volSR}
\title{Calculate the volume of a simple solid of revolution}
\usage{
volSR(intprofile, left = TRUE)
}
\arguments{
\item{intprofile}{An object of class SpatialPolygons* delineating the outline of the left or right-hand side of the interior of a vessel.}

\item{left}{Is it the left or the right-hand side.}
}
\value{
A single numeric value ofr the capacity in the ucubed units of the input.
}
\description{
Given a left or right-hand half of the interior of a simple (no reversions) vessel shape, calculate the likely capacity if it were swept 360* about its middle axis, via the piecewise cylinders methods
}
\examples{
data(dolium)
plot(dolium, col="grey", border=NA, axes=TRUE)
plot(doliumint, border="red", add=TRUE)
volSR(doliumint) ##Input units are metres, cpaacity is 12.8L
}
